+++ model/adjuster/span_id_deduper_test.go
@@ -14,13 +14,7 @@ import (
 	"github.com/jaegertracing/jaeger/model"
 )
 
-var (
-	clientSpanID  = model.NewSpanID(1)
-	anotherSpanID = model.NewSpanID(11)
-	keySpanKind   = "span.kind"
-)
-
-func newTrace() *model.Trace {
+func newDuplicatedSpansTrace() *model.Trace {
 	traceID := model.NewTraceID(0, 42)
 	return &model.Trace{
 		Spans: []*model.Span{
@@ -52,58 +46,8 @@ func newTrace() *model.Trace {
 	}
 }
 
-func TestZipkinSpanIDRenamerTriggered(t *testing.T) {
-	trace := newTrace()
-	deduper := ZipkinSpanIDRenamer()
-	trace, err := deduper.Adjust(trace)
-	require.NoError(t, err)
-
-	clientSpan := trace.Spans[0]
-	assert.Equal(t, clientSpanID, clientSpan.SpanID, "client span ID should not change")
-
-	serverSpan := trace.Spans[1]
-	assert.Equal(t, clientSpanID+1, serverSpan.SpanID, "server span ID should be reassigned")
-	assert.Equal(t, clientSpan.SpanID, serverSpan.ParentSpanID(), "client span should be server span's parent")
-
-	thirdSpan := trace.Spans[2]
-	assert.Equal(t, anotherSpanID, thirdSpan.SpanID, "3rd span ID should not change")
-	assert.Equal(t, serverSpan.SpanID, thirdSpan.ParentSpanID(), "server span should be 3rd span's parent")
-}
-
-func TestZipkinSpanIDRenamerNotTriggered(t *testing.T) {
-	trace := newTrace()
-	trace.Spans = trace.Spans[1:] // remove client span
-
-	deduper := ZipkinSpanIDRenamer()
-	trace, err := deduper.Adjust(trace)
-	require.NoError(t, err)
-
-	serverSpanID := clientSpanID // for better readability
-	serverSpan := trace.Spans[0]
-	assert.Equal(t, serverSpanID, serverSpan.SpanID, "server span ID should be unchanged")
-
-	thirdSpan := trace.Spans[1]
-	assert.Equal(t, anotherSpanID, thirdSpan.SpanID, "3rd span ID should not change")
-	assert.Equal(t, serverSpan.SpanID, thirdSpan.ParentSpanID(), "server span should be 3rd span's parent")
-}
-
-func TestZipkinSpanIDRenamerError(t *testing.T) {
-	trace := newTrace()
-
-	maxID := int64(-1)
-	assert.Equal(t, maxSpanID, model.NewSpanID(uint64(maxID)), "maxSpanID must be 2^64-1")
-
-	deduper := &spanIDDeduper{trace: trace}
-	deduper.groupSpansByID()
-	deduper.maxUsedID = maxSpanID - 1
-	deduper.uniquifyServerSpanIDs()
-	if assert.Len(t, trace.Spans[1].Warnings, 1) {
-		assert.Equal(t, "cannot assign unique span ID, too many spans in the trace", trace.Spans[1].Warnings[0])
-	}
-}
-
 func TestDedupeBySpanID(t *testing.T) {
-	trace := newTrace()
+	trace := newZipkinTrace()
 	deduper := DedupeBySpanID()
 	trace, err := deduper.Adjust(trace)
 	require.NoError(t, err)
